# Cache identifiers for better performance on getIriFromItem