… tests/Bridge/Symfony/Bundle/DependencyInjection/ApiPlatformExtensionTest.php