| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | no
| BC breaks?    | yes
| Deprecations? | no
| Tests pass?   | Not yet (still experimenting)
| Related tickets | #212
| License       | MIT
| Doc PR        | ø

## Add `getIriFromItem` cache

The collection here has ~8k items, no relations. Dev mode active but caching with apcu is forced.

[Blackfire before](https://blackfire.io/profiles/325404d0-c782-4f05-88ed-4676b6c3368b/graph):
- 2.87s spent in `getIriFromItem`
- 2.35 s for `getIdentifiersFromItem`

[Blackfire after](https://blackfire.io/profiles/7215dbe1-efb1-4835-aa8f-5296761c280a/graph):
- 898 ms spent in `getIriFromItem`
- 371 ms for `getIdentifiersFromItem`

When discussing my performance issue with @Simperfit we noticed that `getIdentifiersFromItem` was called a lot (previous method is `getIriFromItem`). This is called for every item in the collection.

### Profiler:

Don't look at the overall time, only the time spent serializing :p.

#### before
![serializer3](https://cloud.githubusercontent.com/assets/1321971/24004191/f030f9aa-0a65-11e7-815c-26b692724071.png)

#### after
![cacheiri](https://cloud.githubusercontent.com/assets/1321971/24053578/caf6203c-0b39-11e7-870e-9c6402ff1fe8.png)


In this PR, we cache the **properties** that are **identifiers** for the given class.