@teohhanhui 2017-03-20
I like the idea, but the implementation is wrong. We should decorate the IriConverter.