@soyuka 2017-03-22
Okay so I did this with a decorator, but because what's slow is `getIdentifiersFromItem` (which is a private method) I'm not sure it's ideal. Indeed, I had to copy the code from `IriConverter` (trying to avoid BC break at the same time), therefore I see no point in decorating :/.

You can check the code to understand what I mean... 

A solution would be to have a new service for `getIdentifiersFromItem` that we can cache without changing the IriConverter. Lmk if you've a better idea.

----- 

From my benchmarks, I know that we can still win a lot by caching the identifiers getters instead of calling the `PropertyAccessor` (`$identifiers[$propertyName] = $item->{$accessor}();`.

Some links (note that I'm using the normalizer dumper, 8000 items collection):

- [IriConverter without cache](https://blackfire.io/profiles/beab0865-7e50-40ed-823c-bc743e69edb3/graph) 8.36s
- [IriConverter with this PR cache](https://blackfire.io/profiles/51b859bd-956e-4f35-b5c5-79d3968717b7/graph) 6.89s
- [IriConverter with direct getter instead of PropertyAccessor + cache](https://blackfire.io/profiles/e9fda268-d964-4692-85cf-c85ed32391c9/graph) 4.78s