@teohhanhui 2017-03-28
@soyuka Okay, since what we're caching is a private method, then decoration indeed makes no sense.

> A solution would be to have a new service for getIdentifiersFromItem that we can cache without changing the IriConverter.

`ItemIdentifiersResolverInterface`?