@soyuka 2017-03-28
`ItemIdentifiersResolverInterface`? Extract is a bit strong no?

Would `get${ucfirst($propertyName)}` be acceptable to retrieve identifiers instead of using the property accessor?