@teohhanhui 2017-03-28
I think "extract" describes better what we're doing here: extracting identifiers from an item.

> Would get${ucfirst($propertyName)} be acceptable to retrieve identifiers instead of using the property accessor?

No, we cannot make such assumptions. Plus, it'd be a BC break.