@dunglas 2017-03-28
What we can do however is provide implementation of the `PropertyAccessorInterface` doing that. I usually do this for perf in my project. Such class should be included directly in Symfony.