@soyuka 2017-03-28
@dunglas You mean having a kind of IdentifierPropertyAccessor witch would end up doing something simpler than the original class (still no assumption could be made regarding `get{$attribute}` ? We also don't need `setValue` in this context.