@soyuka 2017-03-28
Oh I'm doing this already on my project for this particular class :). So, I'll just add a `ItemIdentifiersExtractorInterface` so that identifiers can be cached here but I won't touch the propertyAccessor thing.