@soyuka 2017-03-29
ping @teohhanhui @dunglas 

I tried to avoid breaking changes lmk what you think. 
Note that Behat tests are green, just waiting for implementation approval to add tests (and #1014 comes first obviously).