@soyuka 2017-04-12
I did a great job covering this part as it wasn't tested anywhere (except through IriConverter, but not fully covered). 

There's still a point to discuss: the location of the `IdentifiersExtractor`. 

I'm not sure that it should actually be under `ApiPlatform\Core\Metadata` because it doesn't interfere with metadata, rather uses the metadata to extract identifiers. Therefore, IMHO this class should be in `ApiPlatform\Core\Api` (so is `ResourceClassResolver` for example). Let me know if you disagree we can always move it elsewhere :).